# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AkadiaPay is a school financial platform API for authentication, payment configuration, student billing, payment processing, allocation, and financial ledger.\n\nBusiness process coverage:\n- Authentication and tenant workspace access\n- Payment policy configuration\n- Payment product configuration\n- Student obligation billing\n- Parent/student outstanding bill visibility\n- Payment order processing\n- Payment allocation\n- Financial ledger posting and reporting",
        "title": "AkadiaPay API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "AkadiaPay API Support",
            "email": "support@akadiapay.local"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticates a user and returns JWT-based tenant and role access contexts for the workspaces available to that user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login and get tenant access tokens",
                "parameters": [
                    {
                        "description": "Login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.AuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerAuthLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the active JWT-derived user, role, tenant, and optional student context for the current access token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Get authenticated profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerAuthProfileResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ledger-entries": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists debit and credit accounting entries generated by payment processing and ledger posting. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger Entry"
                ],
                "summary": "List ledger entries",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default 15, max 100)",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by payment order ID",
                        "name": "payment_order_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by account code",
                        "name": "account_code",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter entry date from (YYYY-MM-DD)",
                        "name": "entry_date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter entry date to (YYYY-MM-DD)",
                        "name": "entry_date_to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword for account code, account name, or description",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerLedgerEntryPageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payment-orders/{id}/ledger": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns debit and credit accounting entries generated for one payment order after ledger posting. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger Entry"
                ],
                "summary": "Get ledger entries by payment order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerLedgerEntryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/payment-orders/{id}/post-ledger": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates debit and credit accounting entries for a completed payment order and remains idempotent for repeated post requests. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ledger Entry"
                ],
                "summary": "Post ledger for payment order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SwaggerLedgerEntryListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.AuthLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "domain.AuthLoginResponse": {
            "type": "object",
            "properties": {
                "is_default": {
                    "type": "boolean"
                },
                "role_code": {
                    "type": "string"
                },
                "student_id": {
                    "type": "string"
                },
                "tenant_code": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "tenant_name": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.AuthProfileResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "role_code": {
                    "$ref": "#/definitions/model.RoleCode"
                },
                "tenant_code": {
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "tenant_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation error message"
                }
            }
        },
        "domain.LedgerEntryResponse": {
            "type": "object",
            "properties": {
                "account_code": {
                    "type": "string"
                },
                "account_name": {
                    "type": "string"
                },
                "credit": {
                    "type": "number"
                },
                "debit": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "entry_date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_order_id": {
                    "type": "string"
                }
            }
        },
...
        "domain.SwaggerAuthLoginResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.AuthLoginResponse"
                    }
                }
            }
        },
        "domain.SwaggerAuthProfileResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/domain.AuthProfileResponse"
                }
            }
        },
...
        "domain.SwaggerLedgerEntryListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LedgerEntryResponse"
                    }
                }
            }
        },
        "domain.SwaggerLedgerEntryPageResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.LedgerEntryResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/shared.Pagination"
                }
            }
        },
...
        "model.RoleCode": {
            "type": "string",
            "enum": [
                "SUPER_ADMIN",
                "SCHOOL_ADMIN",
                "TREASURER",
                "TEACHER",
                "PARENT",
                "STUDENT"
            ],
            "x-enum-varnames": [
                "RoleCodeSuperAdmin",
                "RoleCodeSchoolAdmin",
                "RoleCodeTreasurer",
                "RoleCodeTeacher",
                "RoleCodeParent",
                "RoleCodeStudent"
            ]
        },
...
        "shared.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "next_page": {
                    "type": "integer"
                },
                "previous_page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_page": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. Example: \"Bearer eyJhbGciOiJIUzI1NiIs...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}