# internal/payment/handler/ledger_entry_handler.go
package handler

import (
	"akadia/domain"
	"akadia/internal/platform/security"
	"akadia/internal/shared"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/google/uuid"
)

type ledgerEntryHandler struct {
	ledgerEntryService domain.LedgerEntryService
}

func NewLedgerEntryHandler(
	ledgerEntryService domain.LedgerEntryService,
) domain.LedgerEntryHandler {
	return &ledgerEntryHandler{
		ledgerEntryService: ledgerEntryService,
	}
}

// FindAll godoc
// @Summary List ledger entries
// @Description Lists debit and credit accounting entries generated by payment processing and ledger posting. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.
// @Tags Ledger Entry
// @Produce json
// @Param page query int false "Page number (default 1)"
// @Param size query int false "Page size (default 15, max 100)"
// @Param payment_order_id query string false "Filter by payment order ID"
// @Param account_code query string false "Filter by account code"
// @Param entry_date_from query string false "Filter entry date from (YYYY-MM-DD)"
// @Param entry_date_to query string false "Filter entry date to (YYYY-MM-DD)"
// @Param keyword query string false "Search keyword for account code, account name, or description"
// @Success 200 {object} domain.SwaggerLedgerEntryPageResponse
// @Failure 400 {object} domain.ErrorResponse
// @Failure 401 {object} domain.ErrorResponse
// @Failure 403 {object} domain.ErrorResponse
// @Failure 422 {object} domain.ErrorResponse
// @Failure 500 {object} domain.ErrorResponse
// @Security BearerAuth
// @Router /ledger-entries [get]
func (h *ledgerEntryHandler) FindAll(c *gin.Context) {
	authContextValue, _ := c.Get(domain.ContextKeyAuth)
	authContext := authContextValue.(*security.AuthContext)

	pageable, err := shared.GetPageable(c)
	if err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"error": err.Error()})
		return
	}

	var req domain.LedgerEntryFilter
	if err := c.ShouldBindQuery(&req); err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"error": err.Error()})
		return
	}

	page, err := h.ledgerEntryService.FindPaginate(c.Request.Context(), authContext, pageable, &req)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, page)
}

// FindByPaymentOrderID godoc
// @Summary Get ledger entries by payment order
// @Description Returns debit and credit accounting entries generated for one payment order after ledger posting. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.
// @Tags Ledger Entry
// @Produce json
// @Param id path string true "Payment order ID"
// @Success 200 {object} domain.SwaggerLedgerEntryListResponse
// @Failure 400 {object} domain.ErrorResponse
// @Failure 401 {object} domain.ErrorResponse
// @Failure 403 {object} domain.ErrorResponse
// @Failure 404 {object} domain.ErrorResponse
// @Failure 422 {object} domain.ErrorResponse
// @Failure 500 {object} domain.ErrorResponse
// @Security BearerAuth
// @Router /payment-orders/{id}/ledger [get]
func (h *ledgerEntryHandler) FindByPaymentOrderID(c *gin.Context) {
	authContextValue, _ := c.Get(domain.ContextKeyAuth)
	authContext := authContextValue.(*security.AuthContext)

	paymentOrderID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"error": shared.ErrInvalidIDParam.Error()})
		return
	}

	res, err := h.ledgerEntryService.FindByPaymentOrderID(c.Request.Context(), authContext, paymentOrderID)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": res})
}

// PostPayment godoc
// @Summary Post ledger for payment order
// @Description Generates debit and credit accounting entries for a completed payment order and remains idempotent for repeated post requests. Allowed roles: SUPER_ADMIN, SCHOOL_ADMIN, TREASURER.
// @Tags Ledger Entry
// @Produce json
// @Param id path string true "Payment order ID"
// @Success 200 {object} domain.SwaggerLedgerEntryListResponse
// @Failure 400 {object} domain.ErrorResponse
// @Failure 401 {object} domain.ErrorResponse
// @Failure 403 {object} domain.ErrorResponse
// @Failure 404 {object} domain.ErrorResponse
// @Failure 422 {object} domain.ErrorResponse
// @Failure 500 {object} domain.ErrorResponse
// @Security BearerAuth
// @Router /payment-orders/{id}/post-ledger [post]
func (h *ledgerEntryHandler) PostPayment(c *gin.Context) {
	authContextValue, _ := c.Get(domain.ContextKeyAuth)
	authContext := authContextValue.(*security.AuthContext)

	paymentOrderID, err := uuid.Parse(c.Param("id"))
	if err != nil {
		c.JSON(http.StatusUnprocessableEntity, gin.H{"error": shared.ErrInvalidIDParam.Error()})
		return
	}

	if err := h.ledgerEntryService.PostPayment(c.Request.Context(), authContext, paymentOrderID); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	res, err := h.ledgerEntryService.FindByPaymentOrderID(c.Request.Context(), authContext, paymentOrderID)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{"data": res})
}